login

+++ src/app/api/docentes.api.ts
@@ -3,6 +3,8 @@ import {
   DocenteResponse,
 } from "../../interface/docente.interface";
 
+import { auth } from "@/auth";
+
 export async function getAllDocentes(
   offset: number = 0,
   limit: number = 3
@@ -17,8 +19,11 @@ export async function getAllDocentes(
   return response.json();
 }
 
-export async function addDocente(docenteData: DocenteData) {
-  const token = "Agregar tu token de autenticación aquí";
+export async function addDocente(
+  docenteData: DocenteData,
+  token: string | undefined
+) {
+  if (!token) throw new Error("Token no encontrado");
   const res = await fetch("http://localhost:4000/api/v1/docentes", {
     method: "POST",
     headers: {
@@ -31,8 +36,12 @@ export async function addDocente(docenteData: DocenteData) {
   return res.json();
 }
 
-export async function updateDocente(docenteData: DocenteData, id: number) {
-  const token = "Agregar tu token de autenticación aquí";
+export async function updateDocente(
+  docenteData: DocenteData,
+  id: number,
+  token: string | undefined
+) {
+  if (!token) throw new Error("Token no encontrado");
   const res = await fetch(`http://localhost:4000/api/v1/docentes/${id}`, {
     method: "PUT",
     headers: {
@@ -45,8 +54,8 @@ export async function updateDocente(docenteData: DocenteData, id: number) {
   return res.json();
 }
 
-export async function deleteDocente(id: number) {
-  const token = "Agregar tu token de autenticación aquí";
+export async function deleteDocente(id: number, token: string | undefined) {
+  if (!token) throw new Error("Token no encontrado");
   const res = await fetch(`http://localhost:4000/api/v1/docentes/${id}`, {
     method: "DELETE",
     headers: {
